Unify hallway demos and embed textures
