--- src/animations/CircularHallway/README.md
@@ -1,5 +1,0 @@
-# Circular Hallway
-
-This animation shows a simple circular corridor rendered with Three.js. The path slowly curves without any twist, giving the impression of a long loop. A button toggles a Möbius twist so you can switch between the ordinary corridor and the original Möbius version.
-
-No external textures are required.
