+++ src/animations/MobiusWalk/MobiusWalk.tsx
@@ -1,4 +1,4 @@
-import React, { useRef } from 'react';
+import React, { useRef, useState } from 'react';
 import * as THREE from 'three';
 import Canvas3D from '@/components/Canvas3D';
 import { makeCorridorGeometry, DEFAULT_PARAMS, paramToFrame } from './corridorGeometry';
@@ -12,13 +12,15 @@ export interface MobiusWalkProps {
 export default function MobiusWalk({ speed = 2 }: MobiusWalkProps) {
   const camPosT = useRef(0);          // param t along centreline
   const clockRef = useRef(new THREE.Clock());
+  const [twist, setTwist] = useState(true);
 
   const onMount = React.useCallback(({ scene, camera, renderer }: {
     scene: THREE.Scene;
     camera: THREE.PerspectiveCamera;
     renderer: THREE.WebGLRenderer;
   }) => {
-    const geom = makeCorridorGeometry(DEFAULT_PARAMS);
+    const params = { ...DEFAULT_PARAMS, tiltTurns: twist ? 1 : 0 };
+    const geom = makeCorridorGeometry(params);
     const mesh = new THREE.Mesh(geom, corridorMaterial());
     scene.add(mesh);
 
@@ -36,9 +38,9 @@ export default function MobiusWalk({ speed = 2 }: MobiusWalkProps) {
 
     const animate = () => {
       const dt = clockRef.current.getDelta();
-      camPosT.current = (camPosT.current + (speed * dt) / (2 * Math.PI * DEFAULT_PARAMS.radius)) % 1;
+      camPosT.current = (camPosT.current + (speed * dt) / (2 * Math.PI * params.radius)) % 1;
 
-      const { position, quaternion } = paramToFrame(camPosT.current, 0, 0, DEFAULT_PARAMS);
+      const { position, quaternion } = paramToFrame(camPosT.current, 0, 0, params);
       camera.position.copy(position);
       camera.quaternion.copy(quaternion);
 
@@ -56,5 +58,15 @@ export default function MobiusWalk({ speed = 2 }: MobiusWalkProps) {
     animate();
   }, [speed]);
 
-  return <Canvas3D onMount={onMount} />;
+  return (
+    <div>
+      <Canvas3D onMount={onMount} />
+      <button
+        style={{ position: 'absolute', top: 20, left: 20 }}
+        onClick={() => setTwist((t) => !t)}
+      >
+        {twist ? 'Disable Twist' : 'Enable Möbius Twist'}
+      </button>
+    </div>
+  );
 }
