+++ src/index.tsx
@@ -6,7 +6,6 @@ import Correspondence from './animations/Correspondence/Correspondence';
 import ComplexRoots from './animations/ComplexRoots/ComplexRoots';
 import ComplexMultibranch from './animations/ComplexMultibranch/ComplexMultibranch';
 import MobiusWalk from './animations/MobiusWalk/MobiusWalk';
-import CircularHallway from './animations/CircularHallway/CircularHallway';
 
 const routes: Record<string, JSX.Element> = {
   '/': <App />,
@@ -14,8 +13,7 @@ const routes: Record<string, JSX.Element> = {
   '/correspondence': <Correspondence />,
   '/roots': <ComplexRoots />,
   '/multibranch': <ComplexMultibranch />,
-  '/mobius': <MobiusWalk />,
-  '/hallway': <CircularHallway />
+  '/mobius': <MobiusWalk />
 };
 
 function getRoute(): JSX.Element {
